bugfixes for ResNet (plus joining the two files for now)
