+++ ResNet/ResNet50.swift
@@ -193,3 +193,28 @@ public struct ResNet50: Layer {
         return classifier.applied(to: tmp)
     }
 }
+
+func loss(model: ResNet50, images: Tensor<Float>, labels: Tensor<Int32>)
+    -> Tensor<Float> {
+    let logits = model.applied(to: images)
+    let oneHotLabels = Tensor<Float>(
+        oneHotAtIndices: labels, depth: logits.shape[1])
+    return softmaxCrossEntropy(logits: logits, labels: oneHotLabels)
+}
+
+let batchSize: Int32 = 128
+let classCount = 1000
+
+let fakeImageBatch = Tensor<Float>(zeros: [batchSize, 224, 224, 3])
+let fakeLabelBatch = Tensor<Int32>(zeros: [batchSize])
+
+let modeRef = ModeRef()
+var resnet = ResNet50(classCount: classCount, modeRef: modeRef)
+let optimizer = SGD<ResNet50, Float>(learningRate: 0.1, momentum: 0.9)
+
+for _ in 0..<10 {
+    let gradients = gradient(at: resnet) { model in
+        loss(model: model, images: fakeImageBatch, labels: fakeLabelBatch)
+    }
+    optimizer.update(&resnet.allDifferentiableVariables, along: gradients)
+}
